move addComment into comments/actions.js
